Refactor firebasefunctions file
