+++ src/components/includes/CalendarHeader.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Icon } from 'semantic-ui-react';
 
-import { logOut } from '../../firebasefunctions';
+import { logOut } from '../../firebase/user';
 import { useMyCourses } from '../../firehooks';
 
 import { CURRENT_SEMESTER } from '../../constants';
