+++ src/firebase/user.ts
@@ -1,0 +1,58 @@
+import firebase from 'firebase/app';
+
+
+const auth = firebase.auth;
+
+export const userUpload = (user: firebase.User | null, db: firebase.firestore.Firestore) => {
+    if (user != null) {
+        const uid = user.uid;
+        const email = user.email || 'Dummy Email';
+        const displayName = user.displayName || 'Dummy name';
+        const photoUrl = user.photoURL || 'Dummy photo';
+        let stringSplit = -1;
+        let firstName = displayName;
+        let lastName = '';
+        if (displayName != null) {
+            stringSplit = displayName.indexOf(' ');
+            if (stringSplit !== -1) {
+                firstName = displayName.substring(0, stringSplit);
+                lastName = displayName.substring(stringSplit + 1);
+            }
+        }
+        db.runTransaction(async (transaction) => {
+            const userDocumentReference = db.collection('users').doc(uid);
+            const userDocument = await transaction.get(userDocumentReference);
+            if (userDocument.exists) {
+                const partialUserDocument: Partial<FireUser> = {
+                    email,
+                    firstName,
+                    lastName,
+                    photoUrl,
+                };
+                transaction.update(userDocumentReference, partialUserDocument);
+            } else {
+                const fullUserDocument: Omit<FireUser, 'userId'> = {
+                    email,
+                    firstName,
+                    lastName,
+                    photoUrl,
+                    courses: [],
+                    roles: {},
+                };
+                transaction.set(userDocumentReference, fullUserDocument);
+            }
+            // eslint-disable-next-line no-console
+        }).catch(() => console.error('Unable to upload user.'));
+    }
+};
+
+export const logOut = () => {
+    auth()
+        .signOut()
+        .then(() => {
+            // Success
+        })
+        .catch(() => {
+            // Fail
+        });
+};
